refactoring: segregate server http
